backend/requirements.txt
fastapi==0.115.14
fastapi-mcp==0.3.4
uvicorn==0.34.3
gunicorn==23.0.0
granian==2.3.1
rich==14.0.0
requests==2.32.3
sqlmodel==0.0.24
alembic==1.16.2
asyncpg==0.30.0
aiosqlite==0.21.0
colorlog==6.9.0
python-multipart==0.0.20
pydantic==2.11.7
pydantic[settings]==2.11.7
pydantic[email]==2.11.7
psycopg[bynary]==3.2.9
#httpx==0.28.1
openai==1.82.0
groq==0.25.0
python-dotenv==1.1.0
aiosqlite==0.21.0
wheel==0.45.1
setuptools==80.8.0
pillow==11.2.1
google-generativeai==0.8.5
google-genai==1.16.1
passlib==1.7.4
bcrypt==4.3.0 #! for password hashing important this version
python-multipart==0.0.20
joserfc==1.1.0
uuid==1.30
httpx[http2]==0.28.1
hypercorn==0.17.3
uvloop==0.21.0
flask==3.1.1
